inference/pyproject.toml
[tool.poetry]
name = "inference"
version = "0.1.0"
description = "Multimodal RAG Inference Service"
authors = [""]

[tool.poetry.dependencies]
# Python Version
python = ">=3.10, <3.12"

# Web Framework
fastapi = "^0.112.0"
uvicorn = "^0.30.6"

# Data Handling
pydantic = "^2.6.3"
pydantic-settings = "^2.1.0"
numpy = "^1.26.4"
pandas = "^2.2.2"

# Machine Learning & Embeddings
sentence-transformers = "^2.6.1"
open-clip-torch = "^2.24.0"
transformers = "^4.38.0"
torch = {version = ">=2.0.0, <3.0.0", optional = true}

# Vector Database & Search
qdrant-client = "^1.8.0"
langchain = "^0.1.13"

# Database Connections
pymongo = "^4.7.1"

# Image Processing
Pillow = "^10.2.0"
opencv-python-headless = "^4.9.0"

# Logging & Utilities
structlog = "^24.1.0"
rich = "^13.7.1"

[tool.poetry.extras]
ml = ["torch"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"